Refactoring `tedge_config` crate files and directories
Created new directory "tedge_config_cli" to segregate from "system_services"
