+++ crates/common/tedge_config/src/lib.rs
@@ -1,17 +1,8 @@
-mod config_setting;
-mod error;
-mod models;
-mod settings;
-mod tedge_config;
-mod tedge_config_defaults;
-mod tedge_config_dto;
-mod tedge_config_location;
-mod tedge_config_repository;
-
 pub mod system_services;
+pub mod tedge_config_cli;
 
-use self::tedge_config_dto::*;
-pub use self::{config_setting::*, error::*, models::*, settings::*};
-pub use self::{
+use self::tedge_config_cli::tedge_config_dto::*;
+pub use self::tedge_config_cli::{config_setting::*, error::*, models::*, settings::*};
+pub use self::tedge_config_cli::{
     tedge_config::*, tedge_config_defaults::*, tedge_config_location::*, tedge_config_repository::*,
 };
