+++ crates/common/tedge_config/src/tedge_config_cli/models/templates_set.rs
